CSP-101: Add new option to Space below select list on section Layout options
